docker/synth-data-gen/requirements.txt
# Core
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1

# HuggingFace ecosystem
transformers==4.44.0
diffusers==0.30.3
accelerate==0.34.2
huggingface_hub==0.24.6
safetensors==0.4.4
tokenizers==0.19.1

# Image processing
Pillow==10.3.0
opencv-python-headless==4.9.0.80
scikit-image==0.23.2
scipy==1.13.1

# CLIP
open_clip_torch==2.26.1
ftfy==6.2.0

# Background removal
rembg==2.0.57
onnxruntime-gpu==1.18.1

# Math/ML utilities
einops==0.8.0
kornia==0.7.3
numpy==1.26.4

# HTTP + utils
requests==2.32.3
google-generativeai==0.7.2
tqdm==4.66.4
PyYAML==6.0.1
packaging==24.1
